Rename files to match the sole enums they contain
