+++ packages/core/src/nack-errors/digital-channel-errors.ts
@@ -1,5 +1,5 @@
 import { NackError, setPrototypeOf } from "./NackError.js";
-import { NackCode } from "./nack-codes.js";
+import { NackCode } from "./NackCode.js";
 
 export class DigitalChannelNotConfiguredForOutputError extends NackError {
     digitalChannel: number;
