+++ packages/rhsplib/lib/RhspLibErrorCode.ts
