+++ packages/rhsplib/lib/SerialError.ts
